+++ admiral/pkg/clusters/outlierdetection_handler.go
@@ -68,7 +68,7 @@ func (cache *outlierDetectionCache) Delete(identity string, env string) error {
 
 func (od OutlierDetectionHandler) Added(ctx context.Context, obj *v1.OutlierDetection) error {
 	log.Infof(LogFormat, common.Add, common.OutlierDetection, obj.Name, od.ClusterID, common.ReceivedStatus)
-	if common.IsAdmiralStateSyncerMode() {
+	if common.IsAdmiralStateSyncerMode() && common.IsStateSyncerCluster(od.ClusterID) {
 		err := od.RemoteRegistry.RegistryClient.PutCustomData(od.ClusterID, obj.Namespace, obj.Name, common.OutlierDetection, ctx.Value("txId").(string), obj)
 		if err != nil {
 			log.Errorf(LogFormat, common.Add, common.OutlierDetection, obj.Name, od.ClusterID, "failed to put "+common.OutlierDetection+" custom data")
@@ -83,7 +83,7 @@ func (od OutlierDetectionHandler) Added(ctx context.Context, obj *v1.OutlierDete
 
 func (od OutlierDetectionHandler) Updated(ctx context.Context, obj *v1.OutlierDetection) error {
 	log.Infof(LogFormat, common.Update, common.OutlierDetection, obj.Name, od.ClusterID, common.ReceivedStatus)
-	if common.IsAdmiralStateSyncerMode() {
+	if common.IsAdmiralStateSyncerMode() && common.IsStateSyncerCluster(od.ClusterID) {
 		err := od.RemoteRegistry.RegistryClient.PutCustomData(od.ClusterID, obj.Namespace, obj.Name, common.OutlierDetection, ctx.Value("txId").(string), obj)
 		if err != nil {
 			log.Errorf(LogFormat, common.Update, common.OutlierDetection, obj.Name, od.ClusterID, "failed to put "+common.OutlierDetection+" custom data")
@@ -98,7 +98,7 @@ func (od OutlierDetectionHandler) Updated(ctx context.Context, obj *v1.OutlierDe
 
 func (od OutlierDetectionHandler) Deleted(ctx context.Context, obj *v1.OutlierDetection) error {
 	log.Infof(LogFormat, common.Delete, common.OutlierDetection, obj.Name, od.ClusterID, common.ReceivedStatus)
-	if common.IsAdmiralStateSyncerMode() {
+	if common.IsAdmiralStateSyncerMode() && common.IsStateSyncerCluster(od.ClusterID) {
 		err := od.RemoteRegistry.RegistryClient.DeleteCustomData(od.ClusterID, obj.Namespace, obj.Name, common.OutlierDetection, ctx.Value("txId").(string))
 		if err != nil {
 			log.Errorf(LogFormat, common.Delete, common.OutlierDetection, obj.Name, od.ClusterID, "failed to delete "+common.OutlierDetection+" custom data")
